DVD-ROM emulation support

+++ src/io/ide/ata.cc
@@ -84,7 +84,7 @@ ATADeviceFile::~ATADeviceFile()
 {
 }
 
-bool ATADeviceFile::seek(uint32 blockno)
+bool ATADeviceFile::seek(uint64 blockno)
 {
 	sys_fseek(mFile, 512 * (uint64)blockno);
 	return true;
